+++ tools/container/runtime/containerd/containerd_linux.go
@@ -14,7 +14,7 @@
 # limitations under the License.
 **/
 
-package main
+package containerd
 
 import (
 	"fmt"
