+++ tools/container/runtime/containerd/containerd_other.go
@@ -17,7 +17,7 @@
 # limitations under the License.
 **/
 
-package main
+package containerd
 
 import (
 	"errors"
